PT 11203 link up and down arrows to history listbox

See merge request uaa-401-swe/pixel-software/mvgcalc!55

+++ lib/util/persistence.py
@@ -61,7 +61,7 @@ class MvgCalcDataBuffer:
 
     def delete_history_by_display(self):
         try:
-            with open(self.data_store_file_path, 'w'):
+            if os.path.exists(self.data_store_file_path):
                 os.remove(self.data_store_file_path)
         except OSError as e:
             #TODO handle exception
